extract module: js
